Fix test failures in recent bindgen

+++ examples/eg/op0.rs
@@ -38,7 +38,7 @@ pub fn op_extract(perl: &Perl, cv: *const cv, o: *const op) -> Op {
     let cls = perl.op_class(o);
     let oc = unsafe {
         let ty = (*o).op_type();
-        *(&ty as *const u32 as *const opcode)
+        *(&ty as *const u16 as *const opcode)
     };
     match cls {
         OPclass::OPclass_NULL => Op::NULL,
@@ -148,7 +148,7 @@ impl Iterator for OpSiblingIter {
 pub fn op_sibling(op: *const unop) -> *const op {
     // PERL_OP_PARENT is on since 5.26
     if let Some(op) = unsafe {op.as_ref()} {
-        if op.op_moresib() == 1 as u32 {
+        if op.op_moresib() == 1 as u16 {
             op.op_sibparent
         } else {
             std::ptr::null()
